chore(api): using correct methods after segregating
